Add Jazzmin for admin interface customization, update product price field to Decimal, and remove unused login and register templates

+++ shop/models.py
@@ -135,7 +135,7 @@ class Products(models.Model):
     name = models.CharField()
     category = models.ForeignKey(
         Categories, models.DO_NOTHING, db_column="category")
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
     description = models.TextField(blank=True, null=True)
     image = models.CharField(null=True)
     amount_in_stock = models.IntegerField(blank=True, null=True)
